Split post-analysis artifact test helpers
Move shared post-analysis artifact persistence builders into test_support so the behavior-owned test file stays under the oversized test/spec guardrail without a new allowlist exception.
